Reintroduce check order of the k-component candidate after clique heuristic.
This aviods a possible division by zero if the biconnected component of the result
of the clique heuristic is not a k-core in the original graph. It is the last
sanity check that the algorithm does before considering a candidate to be a
k-component. This error was introduced while refactoring.

Also fixed error in reporting results for k = {1, 2}, isolated dyads have to have
connectivity 1, and isolated nodes connectivity 0. The error was also introduced
while refactoring.