[WIP] Fix k_cutset implementation, so far with an (ugly) hack.
Commenting the initial checks for biconnectedness (which return
articulation points insted of running the algorithm) revealed
a problem with the implementation. It was returning some incorrect
node cutsets, in addition of the correct cutsets. Now we only keep
looking for cuts in the condensation graph of the residual network
until we have k cuts. This shoud not be necessary, but I'm still
not sure if I got the whole antichans and closed sets right.

This pass all tests without the biconnectivity shortcut, also added
some new ones. The implementation has to be improved, now the auxiliary
data structures (residual network and auxiliary digraph) are not reused.